+++ ext/tiny_tds/extconf.rb
@@ -35,6 +35,19 @@ DIRS = %w(
   /usr/local
 )
 
+if RbConfig::CONFIG['host_os'] =~ /darwin/i
+  # Ruby below 2.7 seems to label the host CPU on Apple Silicon as aarch64
+  # 2.7 and above print is as ARM64
+  target_host_cpu = Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.7') ? 'aarch64' : 'arm64'
+
+  if RbConfig::CONFIG['host_cpu'] == target_host_cpu
+    # Homebrew on Apple Silicon installs into /opt/hombrew
+    # https://docs.brew.sh/Installation
+    # On Intel Macs, it is /usr/local, so no changes necessary to DIRS
+    DIRS.unshift("/opt/homebrew")
+  end
+end
+
 if ENV["RI_DEVKIT"] && ENV["MINGW_PREFIX"] # RubyInstaller Support
   DIRS.unshift(File.join(ENV["RI_DEVKIT"], ENV["MINGW_PREFIX"]))
 end